travis: working on switching from CMake to Meson.
The good news is that we can use Meson's built-in support for coverage
in conjunction with CodeCov.io to get good coverage reports.  They're
not perfect, but I have a feeling a few artificial and/or never_inline
attributes will help.  Getting CodeCov working was actually the vast
majority of the time I spent on this patch.

The bad news is that, while amd64 is working reliably, pip3 is unable
to install the necessary dependencies on other architectures, and the
versions in Bionic are way too old.  This feels like a bug in one of
the dependencies, so hopefully it will get resolved soon.

I haven't been able to get MSVC working yet either, but TBH I haven't
spent much time on it, and it would probably be a better idea to move
the arm/arm64 MSVC builds over to AppVeyor anyways.

OS X is also not behaving yet either, I haven't looked into it in
detail.

Other than than, we'll lose GCC 4.8 because the version of Python on
that version of Ubuntu is too old for Meson.  PGI is also a problem
since the tests fail in C++ mode and Meson unconditionally builds
the C++ tests.  We could work around this is Meson pretty easily, but
I'm tempted to just drop PGI; it's really buggy, and they don't seem
interested in fixing the many issues I've already reported.

Finally, emscripten is going to require some research to get up and
running.  I know Meson supports it, I'm just not sure how that
support works.